Bug fix: The domDepth calculation is incorrect

+++ src/android/plugin/google/maps/MyPluginLayout.java
@@ -512,7 +512,7 @@ public class MyPluginLayout extends FrameLayout implements ViewTreeObserver.OnSc
         String domIDs[] = HTMLNodes.keySet().toArray(new String[HTMLNodes.size()]);
         Bundle domInfo = HTMLNodes.get(pluginMap.mapDivId);
         RectF htmlElementRect;
-        long mapDivDepth = domInfo.getLong("depth");
+        int mapDivDepth = domInfo.getInt("depth");
 
         for (String domId : domIDs) {
           if (pluginMap.mapDivId.equals(domId)) {
@@ -525,7 +525,7 @@ public class MyPluginLayout extends FrameLayout implements ViewTreeObserver.OnSc
           if (domInfo == null) {
             continue;
           }
-          if (domInfo.getLong("depth") <= mapDivDepth) {
+          if (domInfo.getInt("depth") <= mapDivDepth) {
             continue;
           }
